refactor(headless): split harbor-cell.ts into context-budget, tool-executor, and provider seams (#1084)
Second-pass sink-file split of packages/headless/src/harbor-cell.ts (1810 lines),
a behavior-neutral pure move that keeps `#harbor-cell` / './harbor-cell.js' stable
via re-exports.

- harbor-cell-context-budget-env.ts: the MAKA_CONTEXT_* env compiler
  (HARBOR_CELL_CONTEXT_ENV_KEYS, normalizeHarborCellContextEnv,
  buildHarborCellContextBudgetBackendOptions / *PolicySnapshot /
  TaskLedgerExperimentPolicy and their private helpers + archive record I/O).
  This leaf imports only node:fs/promises, node:crypto, node:path, and
  @maka/runtime TYPES — no runtime backends, child_process, or credential reads.
- harbor-cell-tool-executor.ts: tool executors and subprocess helpers
  (createHarborCellLocalToolExecutor, createHarborHttpToolExecutor,
  buildHarborCellAiSdkTools, secret-stripped child env, shell error helpers).
- provider-env.ts: absorbs provider/credential resolution
  (providerApiKeyEnvName, providerFromEnv, resolveHarborCellAiSdkEnv +
  connection/api-key/credentials.json readers) into the existing seam.
- harbor-cell.ts keeps orchestration + backend wiring (runHarborCell[FromEnv],
  ai-sdk/pi backend registration, Pi CLI env, artifact writes).

runtime-policy-ab-run.ts now imports the env-key list from the light
harbor-cell-context-budget-env.ts, so it no longer transitively pulls
node:child_process, runtime backends, and credentials.json reading.

Dedupe booleanEnv/positiveIntEnv/numericEnv into headless-run-env.ts (the shared
env home, which also now owns the RunHarborCellEnv type) and drop the private
positiveIntEnv copy in harbor-task-runner.ts.
